scripts/appconfig/requirements.txt
azure-identity==1.21.0
azure-mgmt-resource==23.3.0
azure-appconfiguration==1.7.1